# WARNING: head commit changed in the meantime
Merge branch 'merged' of
https://github.com/choongang3000/TeamProject.git into developer_KJ

+++ src/main/java/boardst/controller/BSTDeleteController.java
@@ -40,14 +40,15 @@ public class BSTDeleteController {
 		
 		BSTBean bstbean =bstdao.getBoardByNum(num);
 		
-		if(!bstbean.getImage().equals("")) {
-			File dir = new File(uploadPath,bstbean.getImage());
-			
-			if(dir.exists()) {
-				dir.delete();
+		if(bstbean.getImage() != null) {
+			if(!bstbean.getImage().equals("")) {
+				File dir = new File(uploadPath,bstbean.getImage());
+				
+				if(dir.exists()) {
+					dir.delete();
+				}
 			}
 		}
-
 		int cnt = bstdao.deleteBoard(num);
 		
 		if(cnt>0) {
